requirements.txt
cryptography>=3.4.7
pytest>=6.2.5
Pillow>=9.0.0 